python/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
python-dotenv
dashscope
langchain
langchain-community
httpx
cryptography
pydantic-settings
